Add option to skip server certificate validation
This option will allow gobench to make successful HTTPS requests to
servers that have a self-signed certificate, without the need to
manually import it as a trusted certificate. The flag to control the
new option is named "nc" and it can be set to true to disable the
validation check of the server's certificate. The "nc" flag defaults
to false so server certificate validation is performed by default.